back/pyproject.toml
[project]
name = "comments-back"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django==5.2.2",
    "django-cors-headers>=4.9.0",
    "pillow>=12.1.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "ruff>=0.15.7",
]
